Moved Initialization into MainScene.
So far, the initialization of the singletons (e.g. Notifier) had been
done in the AppDelegate.  This worked fine for iOS, but apparently this
does not work on implementations for other platforms.  I have moved it
into the MainScene class for now, pending the intended solution from
the framework developers.